Fix panic on free error.
The splay tree was fun, but it's more complicated and slow than we
really need. Instead we'll use one extra byte to explicitly store the
containing *Chunk.

Also noteworthy, in an attempt to improve our ability to prove the code
correct we've broken it into self-contained units and thus been able to
reduce the size of some rat's-nest functions:

- A Chunk handles allocations to a contiguous chunk of memory,
  disallowing fragmentation.
- A Chunks handles a contiguous sequence of Chunk objects, also exposing
  an unfragmentable allocation model.
- A Buffer references up to three Chunks objects -- left and right sides
  of a dynamic circular buffer (similarish to a dynamic bip buffer) and
  an overflow region. It dispatches allocations to those Chunks objects
  to preserve the main invariant (ordered by insertion time, right
  elements are before left elements are before overflow elements).
- A CircularAllocator wraps the Buffer interface into an Allocator
  interface.